Move mailjet client into a new package infrastructure/mailing (that will contains mails templates and allow to better test and respect layer segregation)
